Initial uptake of changes - tests passing
