+++ weave/trace_server/interface/base_object_classes/base_object_registry.py
@@ -1,6 +1,7 @@
 from typing import Dict, Type
 
 from weave.trace_server.interface.base_object_classes.base_object_def import BaseObject
+from weave.trace_server.interface.base_object_classes.leaderboard import Leaderboard
 from weave.trace_server.interface.base_object_classes.test_only_example import *
 
 BASE_OBJECT_REGISTRY: Dict[str, Type[BaseObject]] = {}
@@ -18,3 +19,4 @@ def register_base_object(cls: Type[BaseObject]) -> None:
 
 register_base_object(TestOnlyExample)
 register_base_object(TestOnlyNestedBaseObject)
+register_base_object(Leaderboard)
